+++ src/pybritive/commands/clear.py
@@ -1,7 +1,7 @@
 import click
 
-from ..helpers.build_britive import build_britive
-from ..helpers.profile_argument_decorator import click_smart_profile_argument
+from pybritive.helpers.build_britive import build_britive
+from pybritive.helpers.profile_argument_decorator import click_smart_profile_argument
 
 
 @click.group()