+++ src/pybritive/helpers/api_method_argument_decorator.py
@@ -2,6 +2,6 @@ import click
 
 
 def click_smart_api_method_argument(func):
-    from ..completers.api import api_completer
+    from pybritive.completers.api import api_completer
     dec = click.argument('method', shell_complete=api_completer)
     return dec(func)